feat(readme): add project setup instructions and roadmap
The changes in this commit add a new section to the README file that provides
instructions for setting up a new Go project using the `gonew` tool. It also
includes a roadmap of planned features to be added to the project template.
These changes aim to make it easier for developers to get started with this
project and understand the planned future development.